[Silver III] Title: 팩토리얼 나누기, Time: 72 ms, Memory: 11416 KB -BaekjoonHub

+++ 백준/Silver/15996. 팩토리얼 나누기/팩토리얼 나누기.java
@@ -22,15 +22,15 @@ public class Main {
 	public static void main(String[] args) throws IOException {
 		st = new StringTokenizer(br.readLine());
 		
-		long N = Long.parseLong(st.nextToken());
-		long A = Long.parseLong(st.nextToken());
+		int N = Integer.parseInt(st.nextToken());
+		int A = Integer.parseInt(st.nextToken());
 		// 입력완료
 		
 		int answer = 0;
 		
 		// N!에 포함된 소인수 A의 개수 구하기 기본코드
 		for (int i = 1; Math.pow(A, i) <= N; i++) { // A^i가 N보다 커지지 않을 때까지만 반복 (파이썬과 다르게 조건문 설정 가능)
-			answer += N / (int) Math.pow(A, i);
+			answer += N / (int) Math.pow(A, i); // Math.pow()는 double이라 int = int + double 꼴이지만 이클립스가 잡지 못하므로 주의하고 형변환할 것
 		}
 		
 		System.out.println(answer);
